resolucao A18 com float

+++ algoritmo/Atividades/A18_inputDivisao&Round.py
@@ -1,6 +1,6 @@
 # Escreva um programa em Python que peça ao usuário para inserir dois números e calcule a divisão do primeiro número pelo segundo número. Certifique-se de verificar se o segundo número não é zero antes de realizar a divisão. Em seguida, exiba o resultado da divisão.
-n1 = int(input("Insira o primeiro número: "))
-n2 = int(input("Insira o segundo número: "))
+n1 = float(input("Insira o primeiro número: "))
+n2 = float(input("Insira o segundo número: "))
 
 divArredondada = round(n1 / n2)
